// src/renderer/store/types/chatMessages.ts
export enum ChatMessageTypes {
	NEW_MESSAGE_ACTION = "NEW_MESSAGE_ACTION",
	MOD_MESSAGE_ACTION = "MOD_MESSAGE_ACTION",
	SUB_MESSAGE_ACTION = "SUB_MESSAGE_ACTION",
	GIF_SUB_MESSAGE_ACTION = "GIF_SUB_MESSAGE_ACTION",
	SET_SEVEN_TV_EMOTES_ACTION = "SET_SEVEN_TV_EMOTES_ACTION",
	SET_KICK_EMOTES_ACTION = "SET_KICK_EMOTES_ACTION",
	SET_CHANNEL_BADGES_ACTION = "SET_CHANNEL_BADGES_ACTION",
	SET_HOST_ACTION = "SET_HOST_ACTION",
	REMOVE_HOST_ACTION = "REMOVE_HOST_ACTION",
}